jwt intregration

+++ controllers/contactController.js
@@ -6,7 +6,7 @@ const Contact = require("../models/contactModel");
 //@access public
 
 const getAllContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({user_id: req.user.id});
   res.status(200).json(contacts);
 });
 
@@ -23,6 +23,7 @@ const createContacts = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id : req.user.id
   });
   res.status(201).json(contact);
 });
@@ -49,6 +50,11 @@ const updateConatct = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
@@ -65,6 +71,10 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
   }
 
   await Contact.findByIdAndDelete(req.params.id);
